+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - deprecate `deploy` command
 - update go version to 1.22.5
+- `marketplace list-versions` command removed from alpha build
 
 ### BREAKING